Crate bones_lib

The core bones library.

Re-exports

pub use instant;

Modules

ecs

Bones ECS

prelude

Bones lib prelude

reset

Extensions of bones_ecs for utility functions in resetting a World

time

Time functionality for the Bones framework.

Structs

DefaultSessionRunner

The default SessionRunner, which just runs the systems once every time it is run.

Game

The Game encompasses a complete bones game’s logic, independent of the renderer and IO implementations.

GameSystems

A collection of systems associated directly to a Game as opposed to a Session.

Session

A bones game. This includes all of the game worlds, and systems.

SessionBuilder

Builder type used to create Session. If using this directly (as opposed to Sessions::create_with), it is important to rember to finish session and add to Sessions with SessionBuilder::finish_and_add.

SessionOptions

Resource that allows you to configure the current session.

Sessions

Container for multiple game sessions.

Traits

GamePlugin

Trait for plugins that can be installed into a Game.

SessionPlugin

Trait for plugins that can be installed into a Session.

SessionRunner

A session runner is in charge of advancing a Session simulation.

Type Aliases

GameSystem

A system that runs directly on a Game instead of in a specific Session.

SessionCommand

Type of session command