Trait bones_lib::asset::AssetIo

pub trait AssetIo: Sync + Send {
    // Required methods
    fn enumerate_packs(&self) -> Result<Vec<String, Global>, Error>;
    fn load_file(&self, loc: AssetLocRef<'_>) -> Result<Vec<u8, Global>, Error>;
    fn watch(&self) -> Option<Receiver<AssetLoc>>;
}

AssetIo is a trait that is implemented for backends capable of loading all the games assets and returning the raw bytes stored in asset files.

Required Methods

fn enumerate_packs(&self) -> Result<Vec<String, Global>, Error>

List the names of the non-core asset pack folders that are installed.

These names, are not necessarily the names of the pack, but the names of the folders that they are located in. These names can be used to load files from the pack in the load_file() method.

fn load_file(&self, loc: AssetLocRef<'_>) -> Result<Vec<u8, Global>, Error>

Get the binary contents of an asset.

The pack_folder is the name of a folder returned by enumerate_packs(), or None to refer to the core pack.

fn watch(&self) -> Option<Receiver<AssetLoc>>

Subscribe to asset changes.