Struct bones_lib::ecs::components::ComponentStore

#[repr(transparent)]
pub struct ComponentStore<T>where
    T: HasSchema,{ /* private fields */ }

A typed wrapper around UntypedComponentStore.

Implementations

impl<T> ComponentStore<T>where T: HasSchema,

pub fn into_untyped(self) -> UntypedComponentStore

Converts to the internal, untyped ComponentStore.

pub fn from_untyped(untyped: UntypedComponentStore) -> ComponentStore<T>

Creates a ComponentStore from an UntypedComponentStore.

Panics

Panics if the schema doesn’t match T.

pub fn insert(&mut self, entity: Entity, component: T) -> Option<T>

Inserts a component for the given Entity index. Returns the previous component, if any.

pub fn get(&self, entity: Entity) -> Option<&T>

Gets an immutable reference to the component of Entity.

pub fn get_mut(&mut self, entity: Entity) -> Option<&mut T>

Gets a mutable reference to the component of Entity.

pub fn get_many_mut<const N: usize>( &mut self, entities: [Entity; N] ) -> [Option<&mut T>; N]

Get mutable references s to the component data for multiple entities at the same time.

Panics

This will panic if the same entity is specified multiple times. This is invalid because it would mean you would have two mutable references to the same component data at the same time.

pub fn remove(&mut self, entity: Entity) -> Option<T>

Removes the component of Entity. Returns Some(T) if the entity did have the component. Returns None if the entity did not have the component.

pub fn iter(&self) -> impl Iterator<Item = &T>

Iterates immutably over all components of this type. Very fast but doesn’t allow joining with other component types.

pub fn iter_mut(&mut self) -> impl Iterator<Item = &mut T>

Iterates mutably over all components of this type. Very fast but doesn’t allow joining with other component types.

pub fn iter_with_bitset( &self, bitset: Rc<BitSetVec, Global> ) -> Map<UntypedComponentBitsetIterator<'_>, fn(_: SchemaRef<'b>) -> &'b T>

Iterates immutably over the components of this type where bitset indicates the indices of entities. Slower than iter() but allows joining between multiple component types.

pub fn iter_mut_with_bitset( &mut self, bitset: Rc<BitSetVec, Global> ) -> Map<UntypedComponentBitsetIteratorMut<'_>, fn(_: SchemaRefMut<'b, 'b>) -> &'b mut T>

Iterates mutable over the components of this type where bitset indicates the indices of entities. Slower than iter() but allows joining between multiple component types.

pub fn bitset(&self) -> &BitSetVec

Read the bitset containing the list of entites with this component type on it.

pub fn contains(&self, entity: Entity) -> bool

Check whether or not this component store has data for the given entity.