Struct bones_lib::ecs::components::UntypedComponentStore

pub struct UntypedComponentStore { /* private fields */ }

Holds components of a given type indexed by Entity.

We do not check if the given entity is alive here, this should be done using Entities.

Implementations

impl UntypedComponentStore

pub fn new(schema: &'static Schema) -> UntypedComponentStore

Create a arbitrary UntypedComponentStore.

In Rust, you will usually not use UntypedComponentStore and will use the statically typed ComponentStore<T> instead.

pub fn for_type<T>() -> UntypedComponentStore
where T: HasSchema,

Create an UntypedComponentStore that is valid for the given type T.

pub fn schema(&self) -> &'static Schema

Get the schema of the components stored.

pub fn insert_box( &mut self, entity: Entity, data: SchemaBox, ) -> Option<SchemaBox>

Insert component data for the given entity and get the previous component data if present.

Panics

Panics if the schema of T doesn’t match the store.

pub fn try_insert_box( &mut self, entity: Entity, data: SchemaBox, ) -> Result<Option<SchemaBox>, SchemaMismatchError>

Insert component data for the given entity and get the previous component data if present.

Errors

Errors if the schema of T doesn’t match the store.

pub fn insert<T>(&mut self, entity: Entity, data: T) -> Option<T>
where T: HasSchema,

Insert component data for the given entity and get the previous component data if present.

Panics

Panics if the schema of T doesn’t match the store.

pub fn try_insert<T>( &mut self, entity: Entity, data: T, ) -> Result<Option<T>, SchemaMismatchError>
where T: HasSchema,

Insert component data for the given entity and get the previous component data if present.

Errors

Errors if the schema of T doesn’t match the store.

pub unsafe fn insert_raw(&mut self, entity: Entity, data: *mut c_void) -> bool

Returns true if the entity already had a component of this type.

If true is returned, the previous value of the pointer will be written to data.

Safety

The data must be a pointer to memory with the same schema.

If false is returned you must ensure the data pointer is not used after pushing.

pub fn get<T>(&self, entity: Entity) -> Option<&T>
where T: HasSchema,

Get a reference to the component storage for the given Entity.

Panics

Panics if the schema of T doesn’t match.

pub fn try_get<T>( &self, entity: Entity, ) -> Result<Option<&T>, SchemaMismatchError>
where T: HasSchema,

Get a reference to the component storage for the given Entity.

Errors

Errors if the schema of T doesn’t match.

pub fn get_ref(&self, entity: Entity) -> Option<SchemaRef<'_>>

Get a SchemaRef to the component for the given Entity if the entity has this component.

pub fn get_mut<T>(&mut self, entity: Entity) -> Option<&mut T>
where T: HasSchema,

Get a mutable reference to the component storage for the given Entity.

Panics

Panics if the schema of T doesn’t match.

pub fn try_get_mut<T>( &mut self, entity: Entity, ) -> Result<Option<&mut T>, SchemaMismatchError>
where T: HasSchema,

Get a mutable reference to the component storage for the given Entity.

Errors

Errors if the schema of T doesn’t match.

pub fn get_mut_or_insert<T>( &mut self, entity: Entity, f: impl FnOnce() -> T, ) -> &mut T
where T: HasSchema,

Get a mutable reference to component storage for the given Entity if it exists. Otherwise inserts T generated by calling parameter: f.

pub fn get_ref_mut<'a>(&mut self, entity: Entity) -> Option<SchemaRefMut<'a>>

Get a SchemaRefMut to the component for the given Entity

pub fn get_many_mut<const N: usize, T>( &mut self, entities: [Entity; N], ) -> [Option<&mut T>; N]
where T: HasSchema,

Get mutable references s to the component data for multiple entities at the same time.

Panics

This will panic if the same entity is specified multiple times. This is invalid because it would mean you would have two mutable references to the same component data at the same time.

This will also panic if there is a schema mismatch.

pub fn try_get_many_mut<const N: usize, T>( &mut self, entities: [Entity; N], ) -> Result<[Option<&mut T>; N], SchemaMismatchError>
where T: HasSchema,

Get mutable references s to the component data for multiple entities at the same time.

Panics

This will panic if the same entity is specified multiple times. This is invalid because it would mean you would have two mutable references to the same component data at the same time.

Errors

This will error if there is a schema mismatch.

pub fn get_many_ref_mut<const N: usize>( &mut self, entities: [Entity; N], ) -> [Option<SchemaRefMut<'_>>; N]

Get SchemaRefMuts to the component data for multiple entities at the same time.

Panics

This will panic if the same entity is specified multiple times. This is invalid because it would mean you would have two mutable references to the same component data at the same time.

pub fn remove<T>(&mut self, entity: Entity) -> Option<T>
where T: HasSchema,

Remove the component data for the entity if it exists.

Errors

Errors if the schema doesn’t match.

pub fn try_remove<T>( &mut self, entity: Entity, ) -> Result<Option<T>, SchemaMismatchError>
where T: HasSchema,

Remove the component data for the entity if it exists.

Errors

Errors if the schema doesn’t match.

pub fn remove_box(&mut self, entity: Entity) -> Option<SchemaBox>

Remove the component data for the entity if it exists.

pub unsafe fn remove_raw( &mut self, entity: Entity, out: Option<*mut c_void>, ) -> bool

If there is a previous value, true will be returned.

If out is set and true is returned, the previous value will be written to it.

Safety

If set, the out pointer, must not overlap the internal component storage.

pub fn get_single_with_bitset( &self, bitset: Rc<BitSetVec>, ) -> Result<SchemaRef<'_>, QuerySingleError>

Get a reference to the component store if there is exactly one instance of the component.

pub fn get_single_with_bitset_mut( &mut self, bitset: Rc<BitSetVec>, ) -> Result<SchemaRefMut<'_>, QuerySingleError>

Get a mutable reference to the component store if there is exactly one instance of the component.

pub fn iter(&self) -> UntypedComponentStoreIter<'_>

Iterates immutably over all components of this type.

Very fast but doesn’t allow joining with other component types.

pub fn iter_mut(&mut self) -> UntypedComponentStoreIterMut<'_>

Iterates mutably over all components of this type.

Very fast but doesn’t allow joining with other component types.

pub fn iter_with_bitset( &self, bitset: Rc<BitSetVec>, ) -> UntypedComponentBitsetIterator<'_>

Iterates immutably over the components of this type where bitset indicates the indices of entities.

Slower than iter() but allows joining between multiple component types.

pub fn iter_mut_with_bitset( &mut self, bitset: Rc<BitSetVec>, ) -> UntypedComponentBitsetIteratorMut<'_>

Iterates mutable over the components of this type where bitset indicates the indices of entities.

Slower than iter() but allows joining between multiple component types.

pub fn iter_with_bitset_optional( &self, bitset: Rc<BitSetVec>, ) -> UntypedComponentOptionalBitsetIterator<'_>

Iterates immutably over the components of this type where bitset indicates the indices of entities. Iterator provides Option, returning None if there is no component for entity in bitset.

pub fn iter_mut_with_bitset_optional( &mut self, bitset: Rc<BitSetVec>, ) -> UntypedComponentOptionalBitsetIteratorMut<'_>

Iterates mutably over the components of this type where bitset indicates the indices of entities. Iterator provides Option, returning None if there is no component for entity in bitset.

pub fn bitset(&self) -> &BitSetVec

Returns the bitset indicating which entity indices have a component associated to them.

Useful to build conditions between multiple Components’ bitsets.

For example, take two bitsets from two different Components types. Then, bitset1.clone().bit_and(bitset2); And finally, you can use bitset1 in iter_with_bitset and iter_mut_with_bitset. This will iterate over the components of the entity only for entities that have both components.

pub fn into_typed<T>(self) -> ComponentStore<T>
where T: HasSchema,

Convert into a typed ComponentStore.

Panics

Panics if the schema doesn’t match.