Struct bones_lib::ecs::entities::Entities

pub struct Entities { /* private fields */ }

Holds a list of alive entities.

It also holds a list of entities that were recently killed, which allows to remove components of deleted entities at the end of a game frame.

Implementations

impl Entities

pub fn single_with<Q>( &self, query: Q, ) -> (Entity, <<Q as QueryItem>::Iter as Iterator>::Item)
where Q: QueryItem,

Get a single entity and components in the given query if there is exactly one entity matching the query.

Panics

This method panics if the number of matching entities is not exactly one.

pub fn get_single_with<Q>( &self, query: Q, ) -> Result<(Entity, <<Q as QueryItem>::Iter as Iterator>::Item), QuerySingleError>
where Q: QueryItem,

Get a single entity and components in the given query if there is exactly one entity matching the query.

pub fn first_with_bitset(&self, bitset: &BitSetVec) -> Entity

Get the first entity in the given bitset.

Panics

This method panics if there are no entities in the bitset.

pub fn get_first_with_bitset(&self, bitset: &BitSetVec) -> Option<Entity>

Get the first entity in the given bitset.

pub fn first_with<Q>( &self, query: Q, ) -> (Entity, <<Q as QueryItem>::Iter as Iterator>::Item)
where Q: QueryItem,

Get the first entity and components in the given query.

Panics

This method panics if there are no entities that match the query.

pub fn get_first_with<Q>( &self, query: Q, ) -> Option<(Entity, <<Q as QueryItem>::Iter as Iterator>::Item)>
where Q: QueryItem,

Get the first entity and components in the given query.

pub fn iter_with_bitset<'a>( &'a self, bitset: &'a BitSetVec, ) -> EntityIterator<'a>

Iterates over entities using the provided bitset.

pub fn iter_with<Q>( &self, query: Q, ) -> EntitiesIterWith<'_, <Q as QueryItem>::Iter>
where Q: QueryItem,

Iterate over the entities and components in the given query.

The QueryItem trait is automatically implemented for references to Comp and CompMut and for tuples of up to 26 items, so you can join over your mutable or immutable component borrows in your systems.

You can also pass a single component, to iterate only over the components that have alive entities.

Example

fn my_system(entities: Res<Entities>, mut pos: CompMut<Pos>, vel: Comp<Vel>) {
    for (entity, (pos, vel)) in entities.iter_with((&mut pos, &vel)) {
        pos.x += vel.x;
        pos.y += vel.y;
    }
}

You may optionally iterate over components with &Optional(&comp) or mutably with &mut OptionalMut(&mut comp_mut). Entities are not filtered by component in OptionalQueryItem. None is returned for these. If done with single Optional query item, all entities are iterated over.

Syntax is &Optional(&comp), or &mut OptionalMut(&mut comp). Reference to comp and reference to Optional is required for now.

Optional Example

fn my_system(entities: Res<Entities>, mut pos: CompMut<Pos>, vel: Comp<Vel>, pos_max: Comp<PosMax>) {
    for (entity, (pos, vel, pos_max)) in entities.iter_with((&mut pos, &vel, &Optional(&pos_max))) {
        // Update pos from vel on all entities that have pos and vel components
        pos.x += vel.x;
        pos.y += vel.y;

        // limit pos.x by pos_max.x if entity has PosMax component
        if let Some(pos_max) = pos_max {
            if pos.x > pos_max.x {
                pos.x = pos_max.x
            }
        }
    }
}

pub fn create(&mut self) -> Entity

Creates a new Entity and returns it.

This function will not reuse the index of an entity that is still in the killed entities.

pub fn is_alive(&self, entity: Entity) -> bool

Checks if the Entity is still alive.

Returns true if it is alive. Returns false if it has been killed.

pub fn kill(&mut self, entity: Entity)

Kill an entity.

pub fn all_cloned(&self) -> Vec<Entity>

Returns a list of all Entitys cloned into a new vec.

pub fn kill_all(&mut self)

Kills all entities.

pub fn killed(&self) -> &Vec<Entity>

Returns entities in the killed list.

pub fn clear_killed(&mut self)

Clears the killed entity list.

pub fn bitset(&self) -> &BitSetVec

Returns a bitset where each index where the bit is set to 1 indicates the index of an alive entity.

Useful for joining over Entity and ComponentStore<T> at the same time.

pub fn iter(&self) -> EntityIterator<'_>

Iterates over all alive entities.