Crate bones_lib::ecs::prelude::bones_schema::prelude::bones_utils

General utilities for Bones meta-engine crates.

Modules

prelude

The prelude.

Macros

key

Create a new const [Key][key_mod::Key] parsed at compile time.

smallvec

Creates a SmallVec containing the arguments.

Structs

Aligned

Used as a type argument to Ptr, PtrMut and OwningPtr to specify that the pointer is aligned.

Drain

An iterator that removes the items from a SmallVec and yields them by value.

IntoIter

An iterator that consumes a SmallVec and yields its items by value.

LabeledId

A Ulid with a human-readable ascii prefix.

OwningPtr

Type-erased Box-like pointer to some unknown type chosen when constructing this type. Conceptually represents ownership of whatever data is being pointed to and so is responsible for calling its Drop impl. This pointer is not responsible for freeing the memory pointed to by this pointer as it may be pointing to an element in a Vec or to a local in a function etc.

Ptr

Type-erased borrow of some unknown type chosen when constructing this type.

PtrMut

Type-erased mutable borrow of some unknown type chosen when constructing this type.

SmallVec

A Vec-like container that can store a small number of elements inline.

ThinSlicePtr

Conceptually equivalent to &'a [T] but with length information cut out for performance reasons

Unaligned

Used as a type argument to Ptr, PtrMut and OwningPtr to specify that the pointer is not aligned.

Ustr

A handle representing a string in the global string cache.

Enums

CollectionAllocErr

Error type for APIs with fallible heap allocation

LabeledIdCreateError

Error creating a LabeledId.

LabledIdParseError

Errors that can happen while parsing a LabeledId.

MaybeOwned

This type provides a way to store data to which you either have a reference to or which you do own.

MaybeOwnedMut

This type is basically the same as MaybeOwned, but works with mutable references.

Traits

Array

Types that can be used as the backing store for a SmallVec

IsAligned

Trait that is only implemented for Aligned and Unaligned to work around the lack of ability to have const generics of an enum.

PtrExt

Extension trait with utils for Ptr.

PtrMutExt

Extension trait with utils for PtrMut.

ToSmallVec

Convenience trait for constructing a SmallVec

UnsafeCellDeref

Extension trait for helper methods on UnsafeCell

Functions

dangling_with_align

Creates a dangling pointer with specified alignment. See NonNull::dangling.

default

An ergonomic abbreviation for Default::default() to make initializing structs easier. This is especially helpful when combined with “struct update syntax”.

get_short_name

Shortens a type name to remove all module paths.

likely

Hints to the compiler that the branch condition is likely to be true. Returns the value passed to it.

unlikely

Hints to the compiler that the branch condition is unlikely to be true. Returns the value passed to it.

ustr

Create a new Ustr from the given &str.

Type Definitions

HashMap

A HashMap implementing aHash, a high speed keyed hashing algorithm intended for use in in-memory hashmaps.

HashSet

A HashSet implementing aHash, a high speed keyed hashing algorithm intended for use in in-memory hashmaps.

UstrMap

A standard HashMap using Ustr as the key type with a custom Hasher that just uses the precomputed hash for speed instead of calculating it

UstrSet

A standard HashSet using Ustr as the key type with a custom Hasher that just uses the precomputed hash for speed instead of calculating it

Derive Macros

Deref

Derive macro for deriving Deref on structs with one field.

DerefMut

Derive macro for deriving DerefMut on structs with one field.