Struct bones_lib::ecs::prelude::bones_schema::prelude::bones_utils::prelude::OwningPtr

pub struct OwningPtr<'a, A = Aligned>(_, _)
where
    A: IsAligned;

Type-erased Box-like pointer to some unknown type chosen when constructing this type. Conceptually represents ownership of whatever data is being pointed to and so is responsible for calling its Drop impl. This pointer is not responsible for freeing the memory pointed to by this pointer as it may be pointing to an element in a Vec or to a local in a function etc.

This type tries to act “borrow-like” like which means that:

Pointer should be considered exclusive and mutable. It cannot be cloned as this would lead to aliased mutability and potentially use after free bugs.

It must always points to a valid value of whatever the pointee type is.

The lifetime 'a accurately represents how long the pointer is valid for.

Must be sufficiently aligned for the unknown pointee type.

It may be helpful to think of this type as similar to &'a mut ManuallyDrop<dyn Any> but without the metadata and able to point to data that does not correspond to a Rust type.

Implementations

impl<'a> OwningPtr<'a, Aligned>

pub fn to_unaligned(self) -> OwningPtr<'a, Unaligned>

Removes the alignment requirement of this pointer

impl<A> OwningPtr<'_, A>where A: IsAligned,

pub unsafe fn byte_offset(self, count: isize) -> OwningPtr<'_, A>

Calculates the offset from a pointer. As the pointer is type-erased, there is no size information available. The provided count parameter is in raw bytes.

See also: ptr::offset

Safety

The offset cannot make the existing ptr null, or take it out of bounds for its allocation.

If the A type parameter is Aligned then the offset must not make the resulting pointer be unaligned for the pointee type.

pub unsafe fn byte_add(self, count: usize) -> OwningPtr<'_, A>

Calculates the offset from a pointer (convenience for .offset(count as isize)). As the pointer is type-erased, there is no size information available. The provided count parameter is in raw bytes.

See also: ptr::add

Safety

The offset cannot make the existing ptr null, or take it out of bounds for its allocation.

If the A type parameter is Aligned then the offset must not make the resulting pointer be unaligned for the pointee type.

impl<'a> OwningPtr<'a, Aligned>

pub fn make<T, F, R>(val: T, f: F) -> Rwhere F: FnOnce(OwningPtr<'_, Aligned>) -> R,

Consumes a value and creates an OwningPtr to it while ensuring a double drop does not happen.

impl<'a, A> OwningPtr<'a, A>where A: IsAligned,

pub unsafe fn new(inner: NonNull<u8>) -> OwningPtr<'a, A>

Creates a new instance from a raw pointer.

Safety

inner must point to valid value of whatever the pointee type is.

If the A type parameter is Aligned then inner must be sufficiently aligned for the pointee type.

inner must have correct provenance to allow read and writes of the pointee type.

The lifetime 'a must be constrained such that this OwningPtr will stay valid and nothing else can read or mutate the pointee while this OwningPtr is live.

pub unsafe fn read<T>(self) -> T

Consumes the OwningPtr to obtain ownership of the underlying data of type T.

Safety

T must be the erased pointee type for this OwningPtr.

If the type parameter A is Unaligned then this pointer must be sufficiently aligned for the pointee type T.

pub unsafe fn drop_as<T>(self)

Consumes the OwningPtr to drop the underlying data of type T.

Safety

T must be the erased pointee type for this OwningPtr.

If the type parameter A is Unaligned then this pointer must be sufficiently aligned for the pointee type T.

pub fn as_ptr(&self) -> *mut u8

Gets the underlying pointer, erasing the associated lifetime.

If possible, it is strongly encouraged to use the other more type-safe functions over this function.

pub fn as_ref(&self) -> Ptr<'_, A>

Gets an immutable pointer from this owned pointer.

pub fn as_mut(&mut self) -> PtrMut<'_, A>

Gets a mutable pointer from this owned pointer.

impl<'a> OwningPtr<'a, Unaligned>

pub unsafe fn read_unaligned<T>(self) -> T

Consumes the OwningPtr to obtain ownership of the underlying data of type T.

Safety

T must be the erased pointee type for this OwningPtr.