Struct bones_lib::ecs::prelude::bones_schema::prelude::SchemaVec

pub struct SchemaVec { /* private fields */ }

A type-erased Vec-like collection that for items with the same [Schema].

Implementations

impl SchemaVec

pub fn new(schema: &'static Schema) -> SchemaVec

Initialize an empty SchemaVec for items with the given schema.

pub fn try_push<T>(&mut self, item: T) -> Result<(), SchemaMismatchError>
where T: HasSchema,

Push an item to the vec.

Errors

Errors if the schema of T doesn’t match the vec.

pub fn push<T>(&mut self, item: T)
where T: HasSchema,

Push an item to the vec.

Panics

Panics if the schema of T doesn’t match the vec.

pub fn try_push_box( &mut self, item: SchemaBox, ) -> Result<(), SchemaMismatchError>

Push the item into the end of the vector.

pub fn push_box(&mut self, item: SchemaBox)

Push the item into the end of the vector.

pub fn pop_box(&mut self) -> Option<SchemaBox>

Pop the last item off of the end of the vector.

pub fn try_pop<T>(&mut self) -> Result<Option<T>, SchemaMismatchError>
where T: HasSchema,

Pop an item off the vec.

Errors

Errors if the schema of T doesn’t match.

pub fn pop<T>(&mut self) -> Option<T>
where T: HasSchema,

Pop an item off the vec.

Panics

Panics if the schema of T doesn’t match.

pub fn try_get<T>(&self, idx: usize) -> Result<Option<&T>, SchemaMismatchError>
where T: HasSchema,

Get an item in the vec.

Errors

Errors if the schema doesn’t match.

pub fn get<T>(&self, idx: usize) -> Option<&T>
where T: HasSchema,

Get an item in the vec.

Panics

Panics if the schema doesn’t match.

pub fn get_ref(&self, idx: usize) -> Option<SchemaRef<'_>>

Get the item with the given index.

pub fn try_get_mut<T>( &mut self, idx: usize, ) -> Result<Option<&mut T>, SchemaMismatchError>
where T: HasSchema,

Get an item in the vec.

Errors

Errors if the schema doesn’t match.

pub fn get_mut<T>(&mut self, idx: usize) -> Option<&mut T>
where T: HasSchema,

Get an item in the vec.

Panics

Panics if the schema doesn’t match.

pub fn get_ref_mut(&mut self, idx: usize) -> Option<SchemaRefMut<'_>>

Get an item with the given index.

pub fn len(&self) -> usize

Get the number of items in the vector.

pub fn is_empty(&self) -> bool

Returns true if the vector has zero items in it.

pub fn capacity(&self) -> usize

Get the capacity of the backing buffer.

pub fn schema(&self) -> &'static Schema

Get the schema of items in this SchemaVec.

pub fn iter(&self) -> SchemaVecIter<'_>

Iterate over values in the vec

pub fn iter_mut(&mut self) -> SchemaVecIterMut<'_>

Iterate mutably over values in the vec

pub fn into_svec<T>(self) -> SVec<T>
where T: HasSchema,

Convert into a typed SVec.

Panics

Panics if the schema of T doesn’t match this SchemaVec’s schema.

pub fn try_into_svec<T>(self) -> Result<SVec<T>, SchemaMismatchError>
where T: HasSchema,

Try to convert into a typed SVec.

Errors

Errors if the schema of T doesn’t match this SchemaVec’s schema.

pub fn hash(&self) -> u64

Get the hash of this SchemaVec.

Panics

Panics if the inner type doesn’t implement hash.

pub unsafe fn raw_hash(ptr: *const c_void) -> u64

Raw version of the hash() function. Not meant for normal use.

Safety

Pointer must be a valid pointer to a SchemaVec.

pub unsafe fn raw_eq(a: *const c_void, b: *const c_void) -> bool

Raw version of the eq() function. Not meant for normal use.

Safety

Pointers must be valid pointers to SchemaVecs.

pub fn remove(&mut self, index: usize) -> SchemaBox

Remove and return the element at position index within the vector, shifting all elements after it to the left.

Panics

Panics if index is out of bounds.

pub fn clear(&mut self)

Clears the vector, removing all values.

pub fn truncate(&mut self, len: usize)

Shortens the vector, keeping the first len elements and dropping the rest.

If len is greater than the vector’s current length, this has no effect.