Module bones_lib::ecs::prelude::bones_schema::raw_fns

Traits implementing raw function calls for cloning, dropping, and creating default values for Rust types.

Traits

RawClone

Trait implemented automatically for types that implement Clone and can be used to clone the type through raw pointers.

RawDefault

Trait implemented automatically for types that implement Default and can be used to write the default value of the type to a pointer.

RawDrop

Trait implemented automatically for types that implement Drop and can be used to drop the type through a raw pointer.

RawEq

Trait implemented automatically for types that implement Eq that can compare two values through their pointers.

RawHash

Trait implemented automatically for types that implement Hash and can be used compute a u64 hash for a type using a pointer to it.