Struct bones_lib::ecs::prelude::Entity

pub struct Entity(_, _);

An entity index.

They are created using the Entities struct. They are used as indices with Components structs.

Entities are conceptual “things” which possess attributes (Components). As an exemple, a Car (Entity) has a Color (Component), a Position (Component) and a Speed (Component).

Implementations

impl Entity

pub fn new(index: u32, generation: u32) -> Entity

Creates a new Entity from the provided index and generation.

⚠️ Warning: It is not generally recommended to manually create Entitys unless you know exactly what you are doing. This can be useful in certain advanced or unusual use-cases, but usually you should use Entities::create() to spawn entities.

pub fn index(&self) -> u32

Returns the index of this Entity.

In most cases, you do not want to use this directly. However, it can be useful to create caches to improve performances.

pub fn generation(&self) -> u32

Returns the generation of this Entity.

In most cases, you do not want to use this directly. However, it can be useful to create caches to improve performances.