Struct bones_lib::ecs::prelude::SMap

pub struct SMap<K, V>
where
    K: HasSchema,
    V: HasSchema,
{ /* private fields */ }

Typed version of a SchemaMap.

This works essentially like a HashMap, but is compatible with the schema ecosystem.

It is also slightly more efficient to access an SMap compared to a SchemaMap because it doesn’t need to do a runtime schema check every time the map is accessed.

Implementations

impl<K, V> SMap<K, V>
where K: HasSchema, V: HasSchema,

pub fn new() -> SMap<K, V>

Initialize the SMap.

pub fn insert(&mut self, k: K, v: V) -> Option<V>

Insert an entry into the map, returning the previous value, if it exists.

pub fn get(&self, key: &K) -> Option<&V>

Get a reference to an item in the map.

pub fn get_mut(&mut self, key: &K) -> Option<&mut V>

Get a mutable reference to an item in the map.

pub fn remove(&mut self, key: &K) -> Option<V>

Remove an item from the map.

pub fn into_schema_map(self) -> SchemaMap

Convert into an untyped SchemaMap.

pub fn contains_key(&self, key: &K) -> bool

Returns true if the map contains a value for the specified key.

pub fn len(&self) -> usize

Returns the number of elements in the map.

pub fn is_empty(&self) -> bool

Returns true if the map contains no elements.

impl<K, V> SMap<K, V>
where K: HasSchema, V: HasSchema,

pub fn iter( &self, ) -> Map<Iter<'_, SchemaBox, SchemaBox>, for<'a> fn(_: (&'a SchemaBox, &'a SchemaBox)) -> (&'a K, &'a V)>

Iterate over entries in the map.

pub fn iter_mut( &mut self, ) -> Map<IterMut<'_, SchemaBox, SchemaBox>, for<'a> fn(_: (&'a SchemaBox, &'a mut SchemaBox)) -> (&'a K, &'a mut V)>

Iterate over entries in the map.

pub fn keys( &self, ) -> Map<Keys<'_, SchemaBox, SchemaBox>, for<'a> fn(_: &'a SchemaBox) -> &'a K>

Iterate over keys in the map.

pub fn values( &self, ) -> Map<Values<'_, SchemaBox, SchemaBox>, for<'a> fn(_: &'a SchemaBox) -> &'a V>

Iterate over values in the map.

pub fn values_mut( &mut self, ) -> Map<ValuesMut<'_, SchemaBox, SchemaBox>, for<'a> fn(_: &'a mut SchemaBox) -> &'a mut V>

Iterate over values in the map.