bones_lib::ecs::resources

Struct UntypedResource

pub struct UntypedResource { /* private fields */ }

An untyped resource that may be inserted into UntypedResources.

This is fundamentally a Arc<AtomicCell<Option<SchemaBox>>> and thus represents a cell that may or may not contain a resource of it’s schema.

Implementations

impl UntypedResource

pub fn empty(schema: &'static Schema) -> UntypedResource

Initialize a new, empty UntypedResource.

pub fn new(resource: SchemaBox) -> UntypedResource

Creates a new UntypedResource storing the given data.

pub fn from_default(schema: &'static Schema) -> UntypedResource

Create a new UntypedResource for the given schema, initially populated with the default value for the schema.

pub fn clone_data(&self) -> Option<SchemaBox>

Clone the inner data, creating a new copy instead of returning another handle the the same data, as the normal clone() implementation does.

pub fn insert( &self, data: SchemaBox, ) -> Result<Option<SchemaBox>, SchemaMismatchError>

Insert resource data into the cell, returning the previous data.

Errors

Errors if the schema of the data does not match that of this cell.

pub fn remove(&self) -> Option<SchemaBox>

Remove the resource data, returning what was stored in it.

pub fn borrow(&self) -> Ref<'_, Option<SchemaBox>>

Borrow the resource.

pub fn borrow_mut(&self) -> RefMut<'_, Option<SchemaBox>>

Mutably borrow the resource.

pub fn schema(&self) -> &'static Schema

Get the schema of the resource.