Module bones_lib::ecs::stage

Implementation of stage abstraction for running collections of systems over a World.

Structs

CommandQueue

A resource containing the Commands command queue.

Commands

A SystemParam that can be used to schedule systems that will be run at the end of the current SystemStage.

CurrentSystemStage

Resource that is automatically added to the world while a system stage is being run that specifies the unique ID of the stage that being run.

SimpleSystemStage

A collection of systems that will be run in order.

SystemStages

An ordered collection of SystemStages.

Enums

CoreStage

A StageLabel for the 5 core stages.

Traits

StageLabel

Trait for things that may be used to identify a system stage.

SystemStage

Trait for system stages. A stage is a