bones_lib::ecs

Module system

Implements the system API for the ECS.

Structs

In

The system input parameter.

Res

SystemParam for getting read access to a resource.

ResInit

SystemParam for getting read access to a resource and initialzing it if it doesn’t already exist.

ResMut

SystemParam for getting mutable access to a resource.

ResMutInit

SystemParam for getting mutable access to a resource and initializing it if it doesn’t already exist.

StaticSystem

Struct containing a static system.

Traits

IntoSystem

Converts a function into a System.

System

Trait implemented by systems.

SystemParam

Trait used to implement parameters for System functions.

Type Aliases

Comp

SystemParam for getting read access to a ComponentStore.

CompMut

SystemParam for getting mutable access to a ComponentStore.