Module bones_lib::prelude::bones_schema::prelude

The prelude.

Structs

SBox

A typed version of SchemaBox.

SMap

Typed version of a SchemaMap.

SVec

A typed version of a SchemaVec.

SchemaBox

A owning, type-erased Box-like container for types with a [Schema].

SchemaData

Schema information describing the memory layout of a type.

SchemaFieldNotFoundError

Error returned when a field is not found in a schema.

SchemaId

A unique identifier for a schema registered in the SCHEMA_REGISTRY.

SchemaLayoutInfo

Layout information computed for SchemaData.

SchemaMap

Untyped schema-aware “HashMap”.

SchemaMismatchError

Error type when attempting to cast between types with mis-matched schemas.

SchemaRef

An untyped reference that knows the [Schema] of the pointee and that can be cast to a matching type.

SchemaRefMut

An untyped mutable reference that knows the [Schema] of the pointee and that can be cast to a matching type.

SchemaRegistry

A schema registry that alloates SchemaIds for SchemaDatas and returns a registered [&'static Schema][Schema].

SchemaVec

A type-erased Vec-like collection that for items with the same [Schema].

StructFieldInfo

A field in a StructSchemaInfo.

StructSchemaInfo

Schema data for a struct.

Ulid

A Ulid is a unique 128-bit lexicographically sortable identifier

Enums

FieldIdx

The index of a field in a struct in a [Schema].

Primitive

A type of primitive.

SchemaKind

A schema describes the data layout of a type, to enable dynamic access to the type’s data through a pointer.

Statics

SCHEMA_REGISTRY

Global SchemaRegistry used to register SchemaDatas and produce [Schema]s.

Traits

FromType

Trait implemented for types that can produce an instance of themselves from a Rust type.

HasSchema

Trait implemented for types that have a [Schema].

Derive Macros

HasSchema

Derive macro for the HasSchema trait.