Struct bones_lib::prelude::bones_schema::prelude::SchemaBox

pub struct SchemaBox { /* private fields */ }

A owning, type-erased Box-like container for types with a [Schema].

Implementations

impl SchemaBox

pub fn as_ptr(&self) -> *mut c_void

Get a raw pointer to the box data.

pub fn cast_into<T>(self) -> T
where T: HasSchema,

Cast this box to it’s inner type and return it.

Panics

Panics if the schema of the box does not match that of the type you are casting to.

pub fn try_cast_into<T>(self) -> Result<T, SchemaMismatchError>
where T: HasSchema,

Cast this box to it’s inner type and return it.

Errors

Errors if the schema of the box does not match that of the type you are casting to.

pub unsafe fn cast_into_unchecked<T>(self) -> T
where T: HasSchema,

Unsafely convert this box into an owned T.

Safety

The schema of type T must equal that of this box.

pub fn cast_ref<T>(&self) -> &T
where T: HasSchema,

Cast this box to a reference to a type with a representative [Schema].

Panics

Panics if the schema of the box does not match that of the type you are casting to.

pub fn try_cast_ref<T>(&self) -> Result<&T, SchemaMismatchError>
where T: HasSchema,

Cast this box to a reference to a type with a representative [Schema].

Errors

Errors if the schema of the box does not match that of the type you are casting to.

pub fn cast_mut<T>(&mut self) -> &mut T
where T: HasSchema,

Cast this box to a mutable reference to a type with a representing [Schema].

Panics

Panics if the schema of the box does not match that of the type you are casting to.

pub fn try_cast_mut<T>(&mut self) -> Result<&mut T, SchemaMismatchError>
where T: HasSchema,

Cast this box to a mutable reference to a type with a representing [Schema].

Errors

Errors if the schema of the box does not match that of the type you are casting to.

pub fn as_ref(&self) -> SchemaRef<'_>

Borrow this box as a SchemaRef.

pub fn as_mut(&mut self) -> SchemaRefMut<'_>

Borrow this box as a SchemaRefMut.

pub fn new<T>(v: T) -> SchemaBox
where T: HasSchema + Sync + Send,

Create a new SchemaBox from an owned type.

pub unsafe fn uninitialized(schema: &'static Schema) -> SchemaBox

Allocates a SchemaBox for the given [Schema], but doesn’t initialize the memory.

Safety

Accessing the data in an unitinialized SchemaBox is undefined behavior. It is up to the user to initialize the memory pointed at by the box after creating it.

pub fn default(schema: &'static Schema) -> SchemaBox

Create a new SchemaBox for a type with a [Schema] that has a SchemaData::default_fn.

Panics

Panics if the passed in schema doesn’t have a default_fn.

pub fn into_sbox<T>(self) -> SBox<T>
where T: HasSchema,

Convert into an SBox if the schema of T matches.

Panics

Panics if the schema of T doesn’t match that of the box.

pub fn try_into_sbox<T>(self) -> Result<SBox<T>, SchemaBox>
where T: HasSchema,

Convert into an SBox if the schema of T matches.

Errors

Returns an error with the orignal SchemaBox if the schema of T doesn’t match.

pub fn schema(&self) -> &'static Schema

Get the [Schema] for the pointer.

pub unsafe fn from_raw_parts( ptr: NonNull<c_void>, schema: &'static Schema, ) -> SchemaBox

Create a new SchemaBox from raw parts.

This is useful for creating a SchemaBox for data with a schema loaded at runtime and without a Rust type.

Safety

You must insure that the pointer is valid for the given schema.

pub fn forget(self)

Deallocate the memory stored in the box, but don’t run the destructor.

pub fn try_hash(&self) -> Option<u64>

Get the hash of this schema box, if supported.

pub fn hash(&self) -> u64

Get the hash of this schema box.

Panics

Panics if the schema doesn’t implement hash.