Struct bones_lib::prelude::AtomicRef

pub struct AtomicRef<'b, T>where
    T: 'b + ?Sized,{ /* private fields */ }

A wrapper type for an immutably borrowed value from an AtomicRefCell<T>.

Implementations

impl<'b, T> AtomicRef<'b, T>where T: ?Sized,

pub fn clone(orig: &AtomicRef<'b, T>) -> AtomicRef<'b, T>

Copies an AtomicRef.

pub fn map<U, F>(orig: AtomicRef<'b, T>, f: F) -> AtomicRef<'b, U>where F: FnOnce(&T) -> &U, U: ?Sized,

Make a new AtomicRef for a component of the borrowed data.

pub fn filter_map<U, F>( orig: AtomicRef<'b, T>, f: F ) -> Option<AtomicRef<'b, U>>where F: FnOnce(&T) -> Option<&U>, U: ?Sized,

Make a new AtomicRef for an optional component of the borrowed data.