Module bones_lib::render::prelude::prelude

The prelude.

Structs

AssetPath

A path to an asset.

AssetProviderMut

A mutable borrow of an AssetProvider.

AssetProviderRef

A borrow of an AssetProvider.

AssetProviders

A resource that may be used to access AssetProviders for all the different registered asset types.

AssetProvidersResource

The type of the AssetProviders resource.

Handle

A typed handle to an asset.

UntypedHandle

An untyped handle to an asset.

Traits

AssetProvider

Trait for asset providers.

UntypedAssetProvider

Trait implemented for asset providers that can return untyped pointers to their assets.

Type Definitions

ResAssetProviders

Type alias for getting the AssetProviders resource.