bones_lib

Struct GameSystems

pub struct GameSystems {
    pub has_run_startup: bool,
    pub startup: Vec<GameSystem>,
    pub before: Vec<GameSystem>,
    pub after: Vec<GameSystem>,
    pub after_session: HashMap<Ustr, Vec<GameSystem>>,
    pub before_session: HashMap<Ustr, Vec<GameSystem>>,
}

A collection of systems associated directly to a Game as opposed to a Session.

Fields
has_run_startup: bool

Flag which indicates whether or not the startup systems have been run yet.

startup: Vec<GameSystem>

Startup systems.

before: Vec<GameSystem>

Game systems that are run before sessions are run.

after: Vec<GameSystem>

Game systems that are run after sessions are run.

after_session: HashMap<Ustr, Vec<GameSystem>>

Game systems that are run after a specific session is run.

before_session: HashMap<Ustr, Vec<GameSystem>>

Game systems that are run before a specific session is run.

Implementations

impl GameSystems

pub fn add_startup_system<F>(&mut self, system: F) -> &mut Self
where F: FnMut(&mut Game) + Sync + Send + 'static,

Add a system that will run only once, before all of the other non-startup systems.

pub fn add_before_system<F>(&mut self, system: F) -> &mut Self
where F: FnMut(&mut Game) + Sync + Send + 'static,

Add a system that will run on every step, before all of the sessions are run.

pub fn add_after_system<F>(&mut self, system: F) -> &mut Self
where F: FnMut(&mut Game) + Sync + Send + 'static,

Add a system that will run on every step, after all of the sessions are run.

pub fn add_before_session_system<F>( &mut self, session: &str, system: F, ) -> &mut Self
where F: FnMut(&mut Game) + Sync + Send + 'static,

Add a system that will run every time the named session is run, before the session is run.

pub fn add_after_session_system<F>( &mut self, session: &str, system: F, ) -> &mut Self
where F: FnMut(&mut Game) + Sync + Send + 'static,

Add a system that will run every time the named session is run, after the session is run.