Module bones_lib::ecs::prelude

The prelude.

Modules

bones_schema

Simple reflection system based on the #[repr(C)] memory layout.

Macros

bitor

Shorthand for combining bitsets with bit_or.

bitset

Shorthand for setting bits on the bitset container.

impl_bitset

Implements the BitSet trait members for your type through DerefMut.

key

Create a new const [Key][key_mod::Key] parsed at compile time.

smallvec

Creates a SmallVec containing the arguments.

Structs

Aligned

Used as a type argument to Ptr, PtrMut and OwningPtr to specify that the pointer is aligned.

AtomicCell

A mutable memory location with dynamically checked borrow rules This type behaves mostly like core::cell::RefCell. The main difference is that this type uses atomic operations for borrowing. Thus allowing to use it in multi-threaded environment.

AtomicResource

A handle to a resource from a Resources collection.

AtomicSchemaRef

An atomic borrow of a SchemaRef.

AtomicSchemaRefMut

An atomic borrow of a SchemaRefMut.

BitFmt

BitSet formatter.

BitSetVec

The type of bitsets used to track entities in component storages. Mostly used to create caches.

CommandQueue

A resource containing the Commands command queue.

Commands

A SystemParam that can be used to schedule systems that will be run at the end of the current SystemStage.

ComponentStore

A typed wrapper around UntypedComponentStore.

ComponentStores

A collection of ComponentStore<T>.

Drain

An iterator that removes the items from a SmallVec and yields them by value.

Entities

Holds a list of alive entities.

EntitiesIterWith

Iterator over entities returned by Entities::iter_with.

Entity

An entity index.

EntityIterator

Iterator over entities using the provided bitset.

EnumSchemaInfo

Schema data for an enum.

In

The system input parameter.

IntoIter

An iterator that consumes a SmallVec and yields its items by value.

LabeledId

A Ulid with a human-readable ascii prefix.

NotInitialized

An error returned when trying to access an uninitialized component.

OwningPtr

Type-erased Box-like pointer to some unknown type chosen when constructing this type. Conceptually represents ownership of whatever data is being pointed to and so is responsible for calling its Drop impl. This pointer is not responsible for freeing the memory pointed to by this pointer as it may be pointing to an element in a Vec or to a local in a function etc.

Ptr

Type-erased borrow of some unknown type chosen when constructing this type.

PtrMut

Type-erased mutable borrow of some unknown type chosen when constructing this type.

Ref

Wrapper for a borrowed AtomicCell that will released lock on drop.

RefMut

Wrapper for mutably borrowed AtomicCell that will released lock on drop.

RegistryState

The internal state o the SchemaRegistry.

Res

SystemParam for getting read access to a resource.

ResInit

SystemParam for getting read access to a resource and initialzing it if it doesn’t already exist.

ResMut

SystemParam for getting mutable access to a resource.

ResMutInit

SystemParam for getting mutable access to a resource and initializing it if it doesn’t already exist.

Resources

A collection of resources.

SBox

A typed version of SchemaBox.

SMap

Typed version of a SchemaMap.

SVec

A typed version of a SchemaVec.

SchemaBox

A owning, type-erased Box-like container for types with a [Schema].

SchemaData

Schema information describing the memory layout of a type.

SchemaDeserialize

Derivable schema type_data for types that implement Deserialize.

SchemaDeserializer

A struct that implements DeserializeSeed and can be used to deserialize values matching a given [Schema].

SchemaFieldNotFoundError

Error returned when a field is not found in a schema.

SchemaId

A unique identifier for a schema registered in the SCHEMA_REGISTRY.

SchemaLayoutInfo

Layout information computed for SchemaData.

SchemaMap

Untyped schema-aware “HashMap”.

SchemaMismatchError

Error type when attempting to cast between types with mis-matched schemas.

SchemaRef

An untyped reference that knows the [Schema] of the pointee and that can be cast to a matching type.

SchemaRefMut

An untyped mutable reference that knows the [Schema] of the pointee and that can be cast to a matching type.

SchemaRegistry

A schema registry that alloates SchemaIds for SchemaDatas and returns a registered [&'static Schema][Schema].

SchemaSerializer

A struct that implements Serialize and wraps around a SchemaRef to serialize the value using it’s schema.

SchemaVec

A type-erased Vec-like collection that for items with the same [Schema].

SimpleSystemStage

A collection of systems that will be run in order.

SmallVec

A Vec-like container that can store a small number of elements inline.

StaticSystem

Struct containing a static system.

StructFieldInfo

A field in a StructSchemaInfo.

StructSchemaInfo

Schema data for a struct.

SystemStages

An ordered collection of SystemStages.

ThinSlicePtr

Conceptually equivalent to &'a [T] but with length information cut out for performance reasons

Ulid

A Ulid is a unique 128-bit lexicographically sortable identifier

Unaligned

Used as a type argument to Ptr, PtrMut and OwningPtr to specify that the pointer is not aligned.

UntypedAtomicResource

An untyped resource that may be inserted into UntypedResources.

UntypedComponentBitsetIterator

Iterates over components using a provided bitset. Each time the bitset has a 1 in index i, the iterator will fetch data from the storage at index i and return it.

UntypedComponentBitsetIteratorMut

Iterates over components using a provided bitset. Each time the bitset has a 1 in index i, the iterator will fetch data from the storage at index i.

UntypedComponentStore

Holds components of a given type indexed by Entity.

UntypedComponentStoreIter

Mutable iterator over pointers in an untyped component store.

UntypedComponentStoreIterMut

Mutable iterator over pointers in an untyped component store.

UntypedResources

Storage for un-typed resources.

Ustr

A handle representing a string in the global string cache.

VariantInfo

Information about an enum variant for EnumSchemaInfo.

World

The World is simply a collection of Resources, and ComponentStores.

Enums

CollectionAllocErr

Error type for APIs with fallible heap allocation

CoreStage

A StageLabel for the 5 core stages.

EnumTagType

A type for an enum tag for EnumSchemaInfo.

FieldIdx

The index of a field in a struct in a [Schema].

LabeledIdCreateError

Error creating a LabeledId.

LabledIdParseError

Errors that can happen while parsing a LabeledId.

MaybeOwned

This type provides a way to store data to which you either have a reference to or which you do own.

MaybeOwnedMut

This type is basically the same as MaybeOwned, but works with mutable references.

Primitive

A type of primitive.

SchemaKind

A schema describes the data layout of a type, to enable dynamic access to the type’s data through a pointer.

Statics

SCHEMA_REGISTRY

Global SchemaRegistry used to register SchemaDatas and produce [Schema]s.

Traits

Array

Types that can be used as the backing store for a SmallVec

BitSet

The BitSet API.

FromType

Trait implemented for types that can produce an instance of themselves from a Rust type.

FromWorld

Creates an instance of the type this trait is implemented for using data from the supplied World.

HasSchema

Trait implemented for types that have a [Schema].

IntoSystem

Converts a function into a System.

IsAligned

Trait that is only implemented for Aligned and Unaligned to work around the lack of ability to have const generics of an enum.

PtrExt

Extension trait with utils for Ptr.

PtrMutExt

Extension trait with utils for PtrMut.

QueryItem

A type representing a component-joining entity query.

StageLabel

Trait for things that may be used to identify a system stage.

System

Trait implemented by systems.

SystemParam

Trait used to implement parameters for System functions.

SystemStage

Trait for system stages. A stage is a

ToSmallVec

Convenience trait for constructing a SmallVec

UnsafeCellDeref

Extension trait for helper methods on UnsafeCell

UnwrapMany

Helper trait for unwraping each item in an array.

Functions

create_bitset

Creates a bitset big enough to contain the index of each entity. Mostly used to create caches.

dangling_with_align

Creates a dangling pointer with specified alignment. See NonNull::dangling.

default

An ergonomic abbreviation for Default::default() to make initializing structs easier. This is especially helpful when combined with “struct update syntax”.

get_short_name

Shortens a type name to remove all module paths.

likely

Hints to the compiler that the branch condition is likely to be true. Returns the value passed to it.

unlikely

Hints to the compiler that the branch condition is unlikely to be true. Returns the value passed to it.

ustr

Create a new Ustr from the given &str.

Type Definitions

Comp

SystemParam for getting read access to a ComponentStore.

CompMut

SystemParam for getting mutable access to a ComponentStore.

ComponentBitsetIterator

Read-only iterator over components matching a given bitset

ComponentBitsetIteratorMut

Mutable iterator over components matching a given bitset

HashMap

A HashMap implementing aHash, a high speed keyed hashing algorithm intended for use in in-memory hashmaps.

HashSet

A HashSet implementing aHash, a high speed keyed hashing algorithm intended for use in in-memory hashmaps.

UstrMap

A standard HashMap using Ustr as the key type with a custom Hasher that just uses the precomputed hash for speed instead of calculating it

UstrSet

A standard HashSet using Ustr as the key type with a custom Hasher that just uses the precomputed hash for speed instead of calculating it

Derive Macros

Deref

Derive macro for deriving Deref on structs with one field.

DerefMut

Derive macro for deriving DerefMut on structs with one field.

HasSchema

Derive macro for the HasSchema trait.