Struct bones_lib::Session

pub struct Session {
    pub world: World,
    pub stages: SystemStages,
    pub active: bool,
    pub visible: bool,
    pub priority: i32,
    pub runner: Box<dyn SessionRunner>,
}

A bones game. This includes all of the game worlds, and systems.

Fields
world: World

The ECS world for the core.

stages: SystemStages

The system stages.

active: bool

Whether or not this session should have it’s systems run.

visible: bool

Whether or not this session should be rendered.

priority: i32

The priority of this session relative to other sessions in the Game.

runner: Box<dyn SessionRunner>

The session runner to use for this session.

Implementations

impl Session

pub fn new() -> Self

Create an empty Session.

pub fn install_plugin(&mut self, plugin: impl SessionPlugin) -> &mut Self

Install a plugin.

pub fn snapshot(&self) -> World

Snapshot the world state.

This is the same as core.world.clone(), but it is more explicit.

pub fn restore(&mut self, world: &mut World)

Restore the world state.

Re-sets the world state to that of the provided world, which may or may not have been created with snapshot().

This is the same as doing an std::mem::swap on self.world, but it is more explicit.

pub fn set_session_runner(&mut self, runner: Box<dyn SessionRunner>)

Set the session runner for this session.

pub fn reset_internals( &mut self, reset_components: bool, reset_entities: bool, reset_systems: bool, )

Provides an interface for resetting various internal parts of the Session. Note this does not fully reset the entire Session.