Trait bones_lib::SessionRunner

pub trait SessionRunner: Sync + Send + 'static {
    // Required method
    fn step(
        &mut self,
        now: Instant,
        world: &mut World,
        stages: &mut SystemStages
    );
}

A session runner is in charge of advancing a Session simulation.

Required Methods

fn step(&mut self, now: Instant, world: &mut World, stages: &mut SystemStages)

Step the simulation once.

It is the responsibility of the session runner to update the Time resource if necessary.

If no special behavior is desired, the simplest session runner, and the one that is implemented by DefaultSessionRunner is as follows:

fn step(&mut self, now: Instant, world: &mut World, stages: &mut SystemStages) {
    world.resource_mut::<Time>().update_with_instant(now);
    stages.run(world);
}